pyproject.toml
[project]
name = "appels-a-projets"
version = "0.1.0"
description = "Collecte et classification des appels à projets (API + scraping)"
authors = [{name = "Walid", email = "walid@example.com"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.3",
    "python-dateutil>=2.9.0",
    "pyyaml>=6.0.1",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pyairtable>=3.3.0",
    "python-dotenv>=1.2.1",
    "anthropic>=0.75.0",
    "openai>=1.0.0",
    "pypdf>=6.5.0",
    "pdfplumber>=0.11.8",
    "pytesseract>=0.3.10",
    "pdf2image>=1.17.0",
    "pillow>=10.2.0",
    "notion-client>=2.7.0",
]
packages = [{ include = "appels_a_projets" }]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
python-dateutil = "^2.9.0"
pyyaml = "^6.0.1"
pandas = "^2.3.3"
pyairtable = "^3.3.0"
python-dotenv = "^1.2.1"
anthropic = "^0.28.0"
pypdf = "^4.0.0"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "notebook>=7.5.1",
    "ipykernel>=7.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
